rename component to components and googleApi files to thirtdPartyApi files
