+++ client/src/components/Footer.js
@@ -1,7 +1,6 @@
 import {
   Box,
   Container,
-  HStack,
   Stack,
   Text,
   useColorModeValue,
