+++ server/.env.example
@@ -1,3 +1,7 @@
 DATABASE_URL="postgres://localhost:5432/guruprasad"
 GOOGLE_API=GoogleApiKeyGoesHere
 PLACE_ID=GooglePlaceIdGoesHere
+JWKS_URI=""
+AUDIENCE=""
+ISSUER=""
+ALGORITHMS="somestring"
